Filter for completed subjects

+++ 2_extractsubcortical_volumes_version6.sh
@@ -42,14 +42,17 @@ echo ${PWD}
 echo "SubjID,LLatVent,RLatVent,Lthal,Rthal,Lcaud,Rcaud,Lput,Rput,Lpal,Rpal,Lhippo,Rhippo,Lamyg,Ramyg,Laccumb,Raccumb,ICV" > subcortical/LandRvolumes.csv
 for subj_id in `ls -d sub* | grep -v "^subcortical$"`
 do 
-	printf "%s,"  "${subj_id}" >> subcortical/LandRvolumes.csv  #JA
-	for x in Left-Lateral-Ventricle Right-Lateral-Ventricle Left-Thalamus-Proper Right-Thalamus-Proper Left-Caudate Right-Caudate Left-Putamen Right-Putamen Left-Pallidum Right-Pallidum Left-Hippocampus Right-Hippocampus Left-Amygdala Right-Amygdala Left-Accumbens-area Right-Accumbens-area 
-	do
-		printf "%g," `grep  ${x} ${subj_id}/stats/aseg.stats | awk '{print $4}'` >> subcortical/LandRvolumes.csv  #JA
-	done
-	
-	printf "%g" `cat ${subj_id}/stats/aseg.stats | grep IntraCranialVol | awk -F, '{print $4}'` >> subcortical/LandRvolumes.csv  #JA
-	echo "" >> subcortical/LandRvolumes.csv  #JA
+	# only add to report if aseg.stats exists
+	if [ -s "${subj_id}/stats/aseg.stats" ]; then
+		printf "%s,"  "${subj_id}" >> subcortical/LandRvolumes.csv  #JA
+		for x in Left-Lateral-Ventricle Right-Lateral-Ventricle Left-Thalamus-Proper Right-Thalamus-Proper Left-Caudate Right-Caudate Left-Putamen Right-Putamen Left-Pallidum Right-Pallidum Left-Hippocampus Right-Hippocampus Left-Amygdala Right-Amygdala Left-Accumbens-area Right-Accumbens-area 
+		do
+			printf "%g," `grep  ${x} ${subj_id}/stats/aseg.stats | awk '{print $4}'` >> subcortical/LandRvolumes.csv  #JA
+		done
+		
+		printf "%g" `cat ${subj_id}/stats/aseg.stats | grep IntraCranialVol | awk -F, '{print $4}'` >> subcortical/LandRvolumes.csv  #JA
+		echo "" >> subcortical/LandRvolumes.csv  #JA
+	fi
 done
 
 ###############
@@ -77,13 +80,16 @@ read resp
 if [ ${resp} == "y" ]; then
 	for img in `more jnk2.txt`
 	do 
-		echo $img 
-		cd ${enigmadir}/outputs/$img/mri/
-		mri_convert --out_orientation RAS --in_type mgz --out_type nii T1.mgz T1.nii ; 
-		mri_convert --out_orientation RAS --in_type mgz --out_type nii aseg.mgz aseg.nii ; 
-		more $hd/jnk.txt | grep $img
-		fslview T1.nii aseg.nii -t 0.2 -l "MGH-Subcortical"; 
-		rm *.nii
+		# only add to report if aseg.stats exists
+		if [ -s "${x}/stats/aseg.stats" ]; then
+			echo $img 
+			cd ${enigmadir}/outputs/$img/mri/
+			mri_convert --out_orientation RAS --in_type mgz --out_type nii T1.mgz T1.nii ; 
+			mri_convert --out_orientation RAS --in_type mgz --out_type nii aseg.mgz aseg.nii ; 
+			more $hd/jnk.txt | grep $img
+			fslview T1.nii aseg.nii -t 0.2 -l "MGH-Subcortical"; 
+			rm *.nii
+		fi
 	done
 fi
